requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
accelerate=1.0.1=pypi_0
aiohappyeyeballs=2.4.3=pypi_0
aiohttp=3.10.11=pypi_0
aiosignal=1.3.1=pypi_0
asttokens=2.4.1=pypi_0
async-timeout=5.0.1=pypi_0
attrs=24.2.0=pypi_0
av=12.3.0=pypi_0
backcall=0.2.0=pypi_0
black=24.8.0=pypi_0
ca-certificates=2024.9.24=h06a4308_0
certifi=2024.8.30=pypi_0
charset-normalizer=3.4.0=pypi_0
click=8.1.7=pypi_0
comm=0.2.2=pypi_0
datasets=3.1.0=pypi_0
debugpy=1.8.8=pypi_0
decorator=5.1.1=pypi_0
dill=0.3.8=pypi_0
executing=2.1.0=pypi_0
filelock=3.16.1=pypi_0
frozenlist=1.5.0=pypi_0
fsspec=2024.9.0=pypi_0
huggingface-hub=0.26.2=pypi_0
idna=3.10=pypi_0
importlib-metadata=8.5.0=pypi_0
ipykernel=6.29.5=pypi_0
ipython=8.12.3=pypi_0
isort=5.13.2=pypi_0
jedi=0.19.2=pypi_0
jinja2=3.1.4=pypi_0
jupyter-client=8.6.3=pypi_0
jupyter-core=5.7.2=pypi_0
levenshtein=0.25.1=pypi_0
libedit=3.1.20230828=h5eee18b_0
libffi=3.2.1=hf484d3e_1007
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
markupsafe=2.1.5=pypi_0
matplotlib-inline=0.1.7=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.1.0=pypi_0
multiprocess=0.70.16=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pypi_0
networkx=3.1=pypi_0
numpy=1.24.4=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.20.5=pypi_0
nvidia-nvjitlink-cu12=12.6.77=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
openssl=1.1.1w=h7f8727e_0
packaging=24.2=pypi_0
pandas=2.0.3=pypi_0
parso=0.8.4=pypi_0
pathspec=0.12.1=pypi_0
peft=0.13.2=pypi_0
pexpect=4.9.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=10.4.0=pypi_0
pip=24.2=py38h06a4308_0
platformdirs=4.3.6=pypi_0
prompt-toolkit=3.0.48=pypi_0
propcache=0.2.0=pypi_0
psutil=6.1.0=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.3=pypi_0
pyarrow=17.0.0=pypi_0
pyflyby=1.9.3=pypi_0
pygments=2.18.0=pypi_0
python=3.8.0=h0371630_2
python-dateutil=2.9.0.post0=pypi_0
pytz=2024.2=pypi_0
pyyaml=6.0.2=pypi_0
pyzmq=26.2.0=pypi_0
qwen-vl-utils=0.0.8=pypi_0
rapidfuzz=3.9.7=pypi_0
readline=7.0=h7b6447c_5
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
safetensors=0.4.5=pypi_0
setuptools=75.1.0=py38h06a4308_0
six=1.16.0=pypi_0
sqlite=3.33.0=h62c20be_0
stack-data=0.6.3=pypi_0
sympy=1.13.3=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.20.3=pypi_0
toml=0.10.2=pypi_0
tomli=2.1.0=pypi_0
torch=2.4.1=pypi_0
torchvision=0.19.1=pypi_0
tornado=6.4.1=pypi_0
tqdm=4.67.0=pypi_0
traitlets=5.14.3=pypi_0
transformers=4.46.2=pypi_0
triton=3.0.0=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024.2=pypi_0
urllib3=2.2.3=pypi_0
wcwidth=0.2.13=pypi_0
wheel=0.44.0=py38h06a4308_0
xxhash=3.5.0=pypi_0
xz=5.4.6=h5eee18b_1
yarl=1.15.2=pypi_0
zipp=3.20.2=pypi_0
zlib=1.2.13=h5eee18b_1
